third_party/blink/web_tests/platform/mac-mac10.15/virtual/threaded/external/wpt/scroll-animations/scroll-animation-inactive-timeline-expected.txt

This is a testharness.js-based test.
FAIL Play pending task doesn't run when the timeline is inactive. assert_equals: Sanity check the timeline is inactive. expected null but got object "0%"
PASS Animation start and current times are correct if scroll timeline is activated after animation.play call.
PASS Animation start and current times are correct if scroll timeline is activated after setting start time.
PASS Animation current time is correct when the timeline becomes newly inactive and then active again.
Harness: the test ran to completion.
